fix: Budget 관련 객체의 amount 타입 int to long

+++ src/main/java/com/wealdy/saemsembackend/domain/budget/service/BudgetService.java
@@ -44,10 +44,9 @@ public class BudgetService {
     }
 
     public List<GetBudgetDto> getBudgetList(LocalDate date, String userId) {
-        // todo: Budget 의 amount int -> long 으로 바꿀 예정
         User user = userService.getUserById(userId);
         return budgetRepository.findByDateAndUser(date, user).stream()
-            .map(projection -> GetBudgetDto.of(projection.getCategoryName(), projection.getAmount().intValue()))
+            .map(projection -> GetBudgetDto.of(projection.getCategoryName(), projection.getAmount()))
             .toList();
     }
 }
